match types

+++ lib/Conversion/TorchToLinalg/Utils.cpp
@@ -75,7 +75,7 @@ Value torch_to_linalg::getDynamicZeroPaddedTensor(
   Value c0 = b.create<arith::ConstantOp>(loc, b.getI64IntegerAttr(0));
   SmallVector<Value> paddingIncludingUnchanged(unpaddedDims, c0);
   paddingIncludingUnchanged.append(padding);
-  assert(unpaddedDims + padding.size() ==
+  assert(static_cast<int64_t>(unpaddedDims + padding.size()) ==
              cast<RankedTensorType>(input.getType()).getRank() &&
          "sum of unpaddedDims and padding.size() must equal to inputRank");
   for (auto pad = paddingIncludingUnchanged.begin();
